Modify HTML and Script to integrate UI elements
HTML Changes:
Created a main div container for the buttons and custom text message.

Javascript Changes:
Removed commented code.
Added references to various HTML elements for UI modification.
Removed GUI update logic from game() and moved it into a new function
called updateGUIScore().
Modified playRound() to set the winner in a variable rather than
returning it. This variable is passed as parameter to updateGUIScore()
and the function is called.